Configuration settings rework and code reorg
- Removed all IBufferSpecificDictionaryProvider related code.  The new
configuration file implementation makes it irrelevant.  Dictionaries
are global and each configuration can potentially use a different
dictionary based on the configuration.  Similar concept, different
implementation.
- Reduced the content of the VisualStudio.SpellChecker.Definitions
assembly.  For now, it only contains the INaturalTextTag interface
which is useful for creating third-party tagger providers for custom
content types.  An ISpellingService interface is also present but
only contains a single method and may be expanded upon in the future.
- Converted the configuration class to be instance-based rather than
static.  This allows for differing configurations at various levels
within a solution and/or project.
- Moved various classes around and renamed some namespaces to better
reflect their usage.
- Added support for adding project-specific settings files.  This is a
catch-all term that refers to solution, project, folder, and
file-specific settings files.  Files are added using the File | New |
Spell Checker Configuration File for Selected Item or by right clicking
an item in the Solution Explorer and selecting the Add Spell Checker
Configuration File option on the main context menu or the Add sub-menu
(based on the item clicked).
- Converted the configuration dialog into an editor window.  This allows
it to edit the global configuration as well as the project-specific
settings files.  When editing file related settings, properties can be
indicated as inherited.
- Added new File Info user control that contains general info about the
settings file being edited and now the settings will be applied.
- Moved Ignored Character Class setting to the General Settings page.
- Converted the Excluded Extensions property to a hash set and moved it
out onto its own configuration page.
- Implemented creation of configuration settings based on the file being
opened and all applicable settings files that exist.
- Added support for specifying additional dictionary folders.  These
will be searched for dictionaries to make available as a selectable
languge for spell checking and can be solution, project, folder, or
file-specific.
- Reworked the global dictionary class to make use of project-specific
dictionaries.
- Added code to clear the cached global dictionaries when a solution
change is detected.  This is necessary to allow for project-specific
dictionaries to override global dictionaries without them being used
outside of the solution once it is unloaded.
- Added support for determining the dictionary language for localized
resource files based on their filename.  This can be disabled using the
related configuration option.
- All of the above changes close #2.

Other changes made:
- Changed configuration editor Reset button to allow resetting of the
entire configuration or just the selected page.
- Added backtick and ellipsis as valid word break characters.  Fixes #4.